# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation test - First Decision",
        "version": "1.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Lista os usuários do sistema",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de usuários do sistema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userArray"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Cria um usuário no sistema",
                "description": "Cria um usuário na base de dados do sistema",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "description": "Dados do novo item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome do usuário"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste@teste1.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retorna o user criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Cria um usuário no sistema",
                "description": "Cria um usuário na base de dados do sistema",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Filtrar um usuário por ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do novo item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome do usuário Editado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste_editado@teste.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retorna o user criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userObject"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Excluir um usuário do sistema",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Excluir um usuário por ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "retorna uma mensagem de que o item foi deletado com sucesso"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "userBase": {
                "properties": {
                    "id": {
                        "description": "Class User",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "userObject": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/userBase"
                    }
                },
                "type": "object"
            },
            "userArray": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/userBase"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}